refactor(index): own reconciliation state and fenced writer interface
Move exact watermark selection, capability reconciliation phases, the erased
writer interface and its lock-backed implementation, and the existing
incompatible-index open fallback into index. Add six focused regressions.

Keep chain authority, process supervision and the remaining reconciliation
loop in node. No intended schema, consensus or persistence-protocol changes.

Related to #742 and #645. Compiler, formatter, Clippy and test validation
remain pending; published for review as a draft, not merge-ready.
