+++ crates/index/Cargo.toml
@@ -29,11 +29,11 @@ bytemuck.workspace = true
 serde.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
+parking_lot.workspace = true
 
 [dev-dependencies]
 proptest.workspace = true
 serde_json.workspace = true
-parking_lot.workspace = true
 criterion.workspace = true
 tempfile = ">=3.20.0, <4"
 bitcoin.workspace = true
